requirements-test.txt
pytest==7.4.0
pytest-cov==4.1.0
mock==5.1.0
requests-mock==1.11.0 